requirements.txt
django==1.10.1
jsonfield==2.0.2
djangorestframework==3.5.1
django-cors-headers==1.2.2
